ml-backend-with-image/requirements.txt
fastapi
uvicorn[standard]
pydantic
transformers
torch
pillow
requests
imagehash
profanity-check
